+++ sdk/azcore/runtime/pager_test.go
@@ -14,6 +14,7 @@ import (
 	"testing"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/internal/exported"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/internal/mock"
 	"github.com/stretchr/testify/require"
 )
@@ -255,4 +256,79 @@ func TestPagerResponderError(t *testing.T) {
 	page, err := pager.NextPage(context.Background())
 	require.Error(t, err)
 	require.Empty(t, page)
+}
+
+func TestFetcherHelper(t *testing.T) {
+	srv, close := mock.NewServer()
+	defer close()
+	pl := exported.NewPipeline(srv)
+
+	srv.AppendResponse()
+	createReqCalled := false
+	resp, err := FetcherHelper(context.Background(), pl, "", func(ctx context.Context) (*policy.Request, error) {
+		createReqCalled = true
+		return NewRequest(ctx, http.MethodGet, srv.URL())
+	})
+	require.NoError(t, err)
+	require.True(t, createReqCalled)
+	require.NotNil(t, resp)
+	require.EqualValues(t, http.StatusOK, resp.StatusCode)
+
+	srv.AppendResponse()
+	createReqCalled = false
+	resp, err = FetcherHelper(context.Background(), pl, srv.URL(), func(ctx context.Context) (*policy.Request, error) {
+		createReqCalled = true
+		return NewRequest(ctx, http.MethodGet, srv.URL())
+	})
+	require.NoError(t, err)
+	require.False(t, createReqCalled)
+	require.NotNil(t, resp)
+	require.EqualValues(t, http.StatusOK, resp.StatusCode)
+
+	resp, err = FetcherHelper(context.Background(), pl, "", func(ctx context.Context) (*policy.Request, error) {
+		return nil, errors.New("failed")
+	})
+	require.Error(t, err)
+	require.Nil(t, resp)
+
+	srv.AppendError(errors.New("failed"))
+	resp, err = FetcherHelper(context.Background(), pl, "", func(ctx context.Context) (*policy.Request, error) {
+		createReqCalled = true
+		return NewRequest(ctx, http.MethodGet, srv.URL())
+	})
+	require.Error(t, err)
+	require.True(t, createReqCalled)
+	require.Nil(t, resp)
+
+	srv.AppendResponse(mock.WithStatusCode(http.StatusBadRequest), mock.WithBody([]byte(`{ "error": { "code": "InvalidResource", "message": "doesn't exist" } }`)))
+	createReqCalled = false
+	resp, err = FetcherHelper(context.Background(), pl, srv.URL(), func(ctx context.Context) (*policy.Request, error) {
+		createReqCalled = true
+		return NewRequest(ctx, http.MethodGet, srv.URL())
+	})
+	require.Error(t, err)
+	var respErr *exported.ResponseError
+	require.ErrorAs(t, err, &respErr)
+	require.EqualValues(t, "InvalidResource", respErr.ErrorCode)
+	require.False(t, createReqCalled)
+	require.Nil(t, resp)
+}
+
+func TestEncodeNextLink(t *testing.T) {
+	const testURL = "https://contoso.com/"
+	nextLink, err := encodeNextLink(testURL + "query?$skip=5&$filter='foo eq bar'")
+	require.NoError(t, err)
+	require.EqualValues(t, testURL+"query?%24filter=%27foo+eq+bar%27&%24skip=5", nextLink)
+	nextLink, err = encodeNextLink(testURL + "query?%24filter=%27foo+eq+bar%27&%24skip=5")
+	require.NoError(t, err)
+	require.EqualValues(t, testURL+"query?%24filter=%27foo+eq+bar%27&%24skip=5", nextLink)
+	nextLink, err = encodeNextLink(testURL + "query?foo=bar&one=two")
+	require.NoError(t, err)
+	require.EqualValues(t, testURL+"query?foo=bar&one=two", nextLink)
+	nextLink, err = encodeNextLink(testURL)
+	require.NoError(t, err)
+	require.EqualValues(t, testURL, nextLink)
+	nextLink, err = encodeNextLink(testURL + "query?invalid=;semicolon")
+	require.Error(t, err)
+	require.Empty(t, nextLink)
 }