+++ sdk/azcore/runtime/request.go
@@ -38,8 +38,7 @@ const (
 )
 
 // NewRequest creates a new policy.Request with the specified input.
-// It's assumed that the URL's path segments have been properly encoded.
-// Any query parameters will be validated and encoded during construction.
+// The endpoint MUST be properly encoded before calling this function.
 func NewRequest(ctx context.Context, httpMethod string, endpoint string) (*policy.Request, error) {
 	return exported.NewRequest(ctx, httpMethod, endpoint)
 }